Enum synom::IResult

pub enum IResult<I, O> {
    Done(I, O),
    Error,
}

The result of a parser.

Variants
Done(I, O)

Parsing succeeded. The first field contains the rest of the unparsed data and the second field contains the parse result.

Error

Parsing failed.

Methods

impl<'a, O> IResult<&'a str, O>

pub fn expect(self, name: &str) -> O

Unwraps the result, asserting the the parse is complete. Panics with a message based on the given string if the parse failed or is incomplete.

extern crate syn;
#[macro_use] extern crate synom;

use syn::Ty;
use syn::parse::ty;

// One or more Rust types separated by commas.
named!(comma_separated_types -> Vec<Ty>,
    separated_nonempty_list!(punct!(","), ty)
);

fn main() {
    let input = "&str, Map<K, V>, String";

    let parsed = comma_separated_types(input).expect("comma-separated types");

    assert_eq!(parsed.len(), 3);
    println!("{:?}", parsed);
}